requirements.txt
Django==4.0
mysqlclient==2.1.0
requests==2.27.1
djangorestframework==3.13.1
